backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
groq==0.4.2
python-dotenv==1.0.0
requests==2.31.0
reportlab==4.0.7
Pillow>=10.2.0
